// packages/graph/src/types/node.types.ts
export type WorkflowIcon = JSXElementConstructor<any>;

export type WorkflowAction = {
  name: string;
  icon?: string;
  color?: string;
  description?: string;
  supportedBindings?: readonly string[];
};

export type NodeExecutionState =
  "idle" | "running" | "start" | "finish" | "suspended" | "cancelled" | "error";
// ...
type NodeDataTitle =
  | { title: string; i18nTitle?: undefined }
  | { i18nTitle: string; title?: undefined };

export type WorkflowNodeTheme = {
  Icon?: WorkflowIcon;
  icon?: string;
  color?: CSSProperties["color"];
  bgColor?: CSSProperties["color"];
  iconColor?: CSSProperties["color"];
  borderColor?: CSSProperties["color"];
};

export type CommonNodeDadaTypes = NodeDataTitle & {
  stepId?: string;
  taskName?: string;
  ownerDefName?: string;
  ownerBindingKind?: string;
  bindingKind?: string;
  bindingName?: string;
  description?: string;
  groupName?: string;
  stepPath?: FlowStepPath;
  theme: WorkflowNodeTheme;
  level?: number;
  executionState?: NodeExecutionState;
};

type CommonNodeData<T extends ENodeType> = CommonNodeDadaTypes & {
  ports: WorkflowNodePort<T>[];
};

export type BindingSingleData = CommonNodeData<ENodeType.BINDING_SINGLE>;
export type BindingMultiData = CommonNodeData<ENodeType.BINDING_MULTI>;

export type TaskData = CommonNodeData<ENodeType.TASK> & {
  actionName?: string;
};

export type IndicatorData = CommonNodeData<ENodeType.INDICATOR> & {
  indicator?: EIndicatorType;
};

export type BranchPlaceholderData =
  CommonNodeData<ENodeType.BRANCH_PLACEHOLDER> & {
    insertPath?: FlowStepPath;
    onOpenInsertMenu?: OnOpenInsertMenu;
  };
export type BindingPlaceholderData =
  CommonNodeData<ENodeType.BINDING_PLACEHOLDER> & {
    ownerDefName?: string;
    bindingKind?: string;
  };
// ...
export type OperatorData = CommonNodeData<ENodeType.OPERATOR> & {
  operatorType?: EOperatorType;
  strategy?: "wait_all" | "wait_any";
};

export type GroupData = {
  ports: WorkflowNodePort<ENodeType.GROUP>[];
  theme: {
    bgColor: string;
  };
  groupName?: never;
};
// ...
export type NodeDataTypes = {
  [ENodeType.BINDING_SINGLE]: BindingSingleData;
  [ENodeType.BINDING_MULTI]: BindingMultiData;
  [ENodeType.INDICATOR]: IndicatorData;
  [ENodeType.OPERATOR]: OperatorData;
  [ENodeType.TASK]: TaskData;
  [ENodeType.GROUP]: GroupData;
  [ENodeType.BRANCH_PLACEHOLDER]: BranchPlaceholderData;
  [ENodeType.BINDING_PLACEHOLDER]: BindingPlaceholderData;
};

export type GraphNode<T extends keyof NodeDataTypes | null = null> =
  T extends keyof NodeDataTypes
    ? Node<NodeDataTypes[T], T>
    : | Node<BindingSingleData, ENodeType.BINDING_SINGLE>
      | Node<BindingMultiData, ENodeType.BINDING_MULTI>
      | Node<IndicatorData, ENodeType.INDICATOR>
      | Node<OperatorData, ENodeType.OPERATOR>
      | Node<TaskData, ENodeType.TASK>
      | Node<GroupData, ENodeType.GROUP>
      | Node<BranchPlaceholderData, ENodeType.BRANCH_PLACEHOLDER>
      | Node<BindingPlaceholderData, ENodeType.BINDING_PLACEHOLDER>;

export enum ENodeType {
  BINDING_SINGLE = "bindingSingle",
  BINDING_MULTI = "bindingMulti",
  INDICATOR = "indicator",
  OPERATOR = "operator",
  TASK = "task",
  GROUP = "group",
  BRANCH_PLACEHOLDER = "branchPlaceholder",
  BINDING_PLACEHOLDER = "bindingPlaceholder",
}

export enum EEdgeType {
  EDGE_WITH_BUTTON = "edgeWithButton",
}

type FlattenedNodeData<T extends ENodeType = ENodeType> = Omit<
  GraphNode<T>,
  "data"
> &
  GraphNode<T>["data"];

export type IWorkflowNodeContext<T extends ENodeType = ENodeType> = Omit<
  FlattenedNodeData<T>,
  "action"
> &
  Partial<CommonNodeDadaTypes> & {
    action?: WorkflowAction | undefined;
    resolvedTheme: WorkflowNodeTheme & { Icon: WorkflowIcon };
  };

export interface IWorkflowNodeProps {
  node: NodeProps<GraphNode>;
  children: ReactNode;
}